coreutils/kexec: Add support for file-based kexec syscall
Add support for the file-based kexec() syscall. The new -k option
indicates a kernel to load, and -i is the initramfs to use.

Probably this should be factored, and the base kexec support (not
the command) should look at magic numbers and do the right thing.
Save this for later.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>