slashinit: Add a hook to download installer/rescue images
If the mount of the root device fails, download a boot
package from the files specified in the goesinstaller
kernel parameter.

If the target root is not mountable, the 'goesinstaller' parameter specifies
an installer/recovery system to use to repair the system. The parameter to
this is three comma-seperated URLs. The first is mandatory, and is the kernel
to load. The second is the optional initramfs to load. The third is the
optional FDT to load. The kernel is loaded via the kexec command.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>
Signed-off-by: Tom Grennan <tgrennan@platinasystems.com>