liner: Implement more traditional BASH/readline Up/Down behavor
Change behavior of Ctrl/N (Up) and Ctrl/P (Down) to walk through
the history position, one line at a time, like readline (literally just
next and previous line in history).

Add altN and altP to do the incremental search. Note that this isn't
exactly the same as readline, which will match the search string
anywhere. We retain the previous behavior of matching on prefixes.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>