liner: Add SetDuration() to allow for timed input
Add a method to allow for timed input. Calling SetDuration()
on a liner will cause a return of liner.ErrTimeOut if there
is no user input for the specified duration.

The current implementation does not work on Windows or on
non-supported input devices. It also is limited to durations of
25.5 seconds, since it uses the VTIME feature of termios.
These things should probably be fixed before getting this upstream.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>