ls: Improve link handling
Pass around the os.FileInfo returned by os.Stat() and os.ReadDir()
to the lister functions, instead of the filename. Callig os.Stat()
twice causes problems for links, since the subsequent os.Stat()
(on the filename, not on the directory) returns the information on
the link, losing the fact that the file entry was a link. This
simplifies a lot, and lets us not need to add and strip prefixes.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>