goes-coreboot: initial commit of a new machine
Initial commit of a new machine called goes-coreboot.
This machine will have a reduced commmand list, the intended
usage is to be called as a payload by coreboot.

This is commit lays the framework, and is simply a copy
of the example machine.  Follow on commits will flesh out
the detail of the goes-coreboot machine.

Signed-off-by: Mark Pleso <mpleso@platinasystems.com>