shellutils: Initial commit
The shellutils package parses shell input either interactively
or from a script. It takes the input and creates a data structure
consisting of lists of pipelines to execute. The pipeliines are lists
of commands, which are a series of words comprised of tokens.

Finally the command lines are translated into definitions of environment
variables and string argument lists.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>