grub: Initial Commit
This is the initial commit of the Grub Emulator. This consists
of the minimal set of commands used in a Debian/Ubuntu autogenerated
/boot/grub/grub.cfg file.

There are many more useful grub commands, but they are not included
if they are not used by Debian or Ubuntu. This will change in the
future.

Commands that are NOP:
	 background_color
	 clear
	 export
	 gfxmode
	 insmod
	 loadfont
	 recordfail
	 search
	 terminal_output

Commands implemented:
	 initrd
	 linux
	 menuentry
	 *submenu

* submenu is parsed properly, but all menus are placed into a single
menu list.

This is a basic proof of concept of grub.cfg parsing.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>