goes: Rework conditionals as blocks using recursion
Rework parser/CLI/goes to support block-type statements in
a general fashion using recursion and closures.

The lexer and executor now work in two phases. Commands are parsed
to return a function, which, when invoked, executes the parsed commands.

Block commands work by returning a function which executes all of the
blocks. Commands like if...then...else return a function which executes
one or the other function.

So, for example, if...then...else works by running the "if" closure,
and then choosing to run the "then" or "else" closure based on exit
status.

	runfun := func(stdin io.Reader, stdout io.Writer, stderr io.Writer, isFirst bool, isLast bool) error {
		err := runList(ifList, stdin, stdout, stderr)
		if err == nil {
			err = runList(thenList, stdin, stdout, stderr)
		} else {
			err = runList(elseList, stdin, stdout, stderr)
		}
		return err
	}

Simple commands just return a function which executes the command.
Complex commands are functions that read the parsers input tokens
and execute the subcommands.

The high level CLI takes these functions and connects them via pipes,
as well as handling compound operators && and ||.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>