goes: Give daemon goroutines 500 milliseconds to finish on SIGTERM
When we receive SIGTERM, we call the Close() method on the underlying
command if there is one. Most commands close a channel at this point,
causing a blocked goroutine to run to do cleanup. However, immediately
after doing this, goes closes the I/O channels and exits, which
does not allow for the cleanup to happen.

Wait 500 milliseconds after signalling a Close() method to allow
other goroutines to run and perform the actual close operation.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>