stop: Make DontFork so we don't kill our parents
Make into a DontFork command, so that we do not kill our parent
process and turn ourself into a zombie.

The kill operation only avoids itself and PID 1. It does
not check to avoid killing its PPID, so rather than adding that
complexity, simply make the command DontFork so that the cli is
not killed.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>