rework tests
This started as an integration of xeth and vnet tests.

Through this is wanted a -test.dryrun to list available tests
(-test.list doesn't show subtests)

I was able to do that with an interface but in doing so, I then
saw how to share a lot of the code in the docker tests with
a struct wrapper. This then lead to cleanup of the docker commands.

I think all of the non-stress tests individually although some multi
suite patterns fail. For example, these pass:

$ sudo ./goes-platina-mk1.test -test.run=Test/vnet/docker/frr/bgp/eth -test.v
$ sudo ./goes-platina-mk1.test -test.run=Test/vnet/docker/frr/isis/eth -test.v
$ sudo ./goes-platina-mk1.test -test.run=Test/vnet/docker/frr/ospf/eth -test.v

But this fails a isis connectivity test:

$ sudo ./goes-platina-mk1.test -test.run=Test/vnet/docker/frr/.*/eth -test.v

To see available tests:

$ ./goes-platina-mk1.test -test.dryrun

Note that the docker tests are now names Test/vnet/docker/.*

Signed-off-by: Tom Grennan <tgrennan@platinasystems.com>