scp: Fix overly restrictive check on directories
Turns out the client doesn't always send -r or -d when the target
is a directory (in particular, it likes to send "." as the target
directory on the command line). So allow a target to be a directory
even if the client didn't tell us it expected it to be.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>