goes: Don't fork when running tests
The golang test infrastructure builds itself into the target being
tested. It generates a test file with a binary name ending with
".test". If you execute this (via /proc/self/exe) it confuses the
test framework. To faciliate tests which exercise the goes infrastructure,
detect that we are running in the test framework and run commands
directly instead of forking ourselves.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>