nop_test: Fix error formatting from when main fails
Who is testing the test that is testing me?

Although we got 100% test coverage on the tests, the coverage analyzer
does not seem to include the tests itself in coverage analysis. As
a result, the test of Main (which does not ever return errors) had
a bug in the formatting of error results. In particular, I had swapped
two arguments, and also used the error wrap (%w) operator which does
not appear to work properly in t.Errorf() (it is a new feature in
fmt.Errorf().

I found this when I used the 'nop' test to be a template for another
package, which does return errors from its Main() method.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>