Jenkinsfile: Now that go builds anywere, don't override checkout directory
This cleans up a lot of things. In the pre-module days, we had to build
in a specific directory, so we had to re-checkout the top of tree. This
means that we needed a different Jenkinsfile for each branch. This also
broke checking of GitHub pull requests.

Now build in the default Jenkins project workspace.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>