goes: Fix error handling (found by grub menu tests)
Return errors from Block() functions instead of ignoring them.

Use error wrapping when propagating errors from commands. This allows
for the use of errors.Is() to check for wrapped errors.

Propogate io.EOF up the error stack

Commands need to be able to return io.EOF. The CLI needs to mask return
io.EOF (which it does), but not the actual command run by the CLI.

Fix error return in MakeListFunc()

Errors were dropped by MakeListFunc(). They should be propogated to the
caller.