redisd: Support dynamic IP address changes
Now that the BMC has IP addresses set by GoES (as opposed to using
kernel autoconfiguration), the management interface may come up after
redis starts.

Track the addresses we listen on, and remove servers dynamically if
the address goes away.

Look for new addresses to appear, and listen on them as they are added.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>