dhcpcd: Fix DHCP renew and crashes with no address offer
Don't crash if we receive an offer with no IP address. Instead, ignore
the offer and try again.

If a renew fails, try getting a new lease.

Instead of renewing once at at lease expiry, first attempt to renew
a lease at half of the lease expiration. If this fails, attempt again
at half the remaining time, and keep doing this until the lease
expires.

When the lease expires, attempt to get a new lease.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>