install: Use resolvconf to set up DNS
It becomes standard in future versions of Debian, so might as well
move to using resolvlconf for DNS setup now.

When setting up the chroot, copy the configured /etc/resolv.conf into
the target filesystem. Then when resolvconf is installed, it'll
igrate this into the proper location and replace /etc/resolv.conf
with its version. Previous Debian versions copied /etc/resolv.conf but
this doesn't happen with resolvconf.

Enable system-resolved via systemctl when in the chroot. This will
cause resolved to be enabled in the target.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>