shellutils: Add glob parsing
Add support for globs in input commands.

If a pattern being parsed contains glob characters `*' `?' or `[',
the rest of the input until whitespace is added to the command line
as a single token containing the entire glob character.

(*Cmdline) Slice() will expand globs transparently, and return an
expanded string.

A new method (*Word) Expand() will expand a word into a slice of strings
doing glob expansion. If the word does not contain globs, it'll return
a slice of length one.

The (*Word) String() method will return the glob as a literal string,
with no expansion.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>