vpn: make file flags relative to CWD or config dir
If these file flag arguments don't contain the respective
file-system separator, slash or back-slash, look for the
named file w/in the current then CONFIG directory.

Also, if available, overwrite flag defaults environment.

  -admins	PREFIX_ADMINS
  -cert		PREFIX_CERT
  -exchanges	PREFIX_EXCHANGES
  -hosts	PREFIX_HOSTS
  -registry	PREFIX_REGISTRY
  -sig		PREFIX_SIG

Where PREFIX is associated with the main package name as

	goes	GOES_VPN
	go-vpn	GO_VPN

In other words, hyphen substitute of the to-upper main
with, if necessary, an attached (_VPN) suffix.

Signed-off-by: Tom Grennan <tgrennan@platinasystems.com>